Use magic mock to mock handler object
New attributes and processes were added to the callback wrapper
which are dependent on `self.batch_size` and a couple of other
attributes/messages of the Handler class. As such we now need to use
magic mock instead of mock for the handler object passed as reference.
Otherwise handler object throws and exception which interferes with
tests.