Do not use @hybrid_property in paths
The path artist_links.artist.gid is not being processed correctly
by iterate_path_values instead of returning the gid, it returns the
artist. LinkArtistUrl.artist is a @hybrid_property and the check
isinstance(LinkArtistUrl.artist, InstrumentedAttribute) returns false.
To identify hybrid_attributes in iterate_path_values, we could use
https://docs.sqlalchemy.org/en/14/orm/internals.html#sqlalchemy.orm.InspectionAttr.extension_type .
I tried this, but currently it does not seem to work. Hence, it is
better that we do not use hybrid attributes in paths for now.