Setup real database for tests
The test config should be copied inside the Dockerfile manually. Copying
config.test.ini to config.ini in a local or jenkins script before building
the docker image will not work because config.ini is included in .dockerignore.
Therefore, config.ini will be ignored and not be copied to the image otherwise.

Co-authored-by: yvanzo <yvanzo@users.noreply.github.com>