SEARCH-653 (I): Drop if exists or continue
When dropping functions and triggers from an SQL database, it would fail
if some function or trigger is missing.  It happens when these functions
and triggers have been created using a previous version of SIR.

This patch prevents the scripts that drop functions and triggers that
may not exist to fail.  This is a first step towards allowing new
versions of SIR to drop functions and triggers created using old ones.

Run `./GenerateDropSql.pl` to regenerate `sql/Drop*.sql` files