Remove volume mount from docker-compose.test.yml
This erases all the code from the image built in CI. So remove the mount,
the downside is that we need to build the image for tests each time. An
alternative could be to have separate docker compose files for CI and local
tests, perhaps even using one as overlay over the other.