Fix mocking of entity.build_entity_query
The build_entity_query method of all entities is mocked in test_index_by_fk_*
tests. These mocks are never reset after the test is over. This causes
the mock to be visible instead of the actual method in other tests causing
those to fail. Fix the issue by using mock.patch and cleaning up after
test run is over.