Optionally accept a session parameter in indexing functions
We want to these indexing functions with real database tests. Thus,
accept an optional session parameter in these functions. In tests,
we create a session, start a transaction, insert data from sql files,
query the data in the same transaction and at the end of the test
rollback the transaction. As a result, we are able to get away without
resetting the database in between the tests.

If we don't pass the same session here, then it may not be possible to
use this method because changes in a transaction are not visible outside
it.

This code earlier used to be inside a with but the queries executed here
are only select, so we can do without the `with` here as well.