Fix overlapping relationship SAWarning in custom models
The test logs are filled with warnings like:

SAWarning: relationship 'AreaTag.area' will copy column area.id to column
area_tag.area, which conflicts with relationship(s): 'CustomArea.tags'
(copies area.id to area_tag.area). If this is not the intention, consider
if these relationships should be linked with back_populates, or if
viewonly=True should be applied to one or more if they are read-only. For
the less common case that foreign key constraints are partially overlapping,
the orm.foreign() annotation can be used to isolate the columns that should
be written towards. To silence this warning, add the parameter
'overlaps="tags"' to the 'AreaTag.area' relationship.

This particular warnings arises from `tags = relationship("AreaTag")` in
CustomArea. The reason (AFAIU) is that `tags` relationship links `Area`
to `AreaTag` and `area` in `AreaTag` links AreaTag back to `Area`.
However, these relationships aren't linked with a backref so it isn't
known to SQLAlchemy that both `relationship` properties represent
the same relation. Possible fixes include using `back_ref` to make
it known to SQLAlchemy that both relationships are same. That approach
however needs changes in mbdata.

In many cases like that of `Tag` and `Alias` relationships, it seems
like a good idea to update mbdata and say add `tags` and `aliases`
attributes directly to `Area` model and other entities. But there
are a few cases like `artist_links` and `recording_links` in
`CustomWork` which probably don't make sense to be added to mbdata
and should continue to live in custom sir models.

Another alternative approach to get rid of the warning as the message
suggests is setting `viewonly=True`, since we only ever read the data
in sir and never modify this option works fine for us. Updating mbdata
likely needs to consider implications on other users of the library so
currently I am not inclined to pursue the approach. Hence, going with
the `viewonly=True` approach.