Wrap filter in list when modifying underlying list
In a few cases, we modify the underlying list while iterating on a
filter of the same list. filter is lazy evaluated so modifying the
underlying list causes elements to go missing from the filter. To avoid
this, wrap filter in list() to complete evaluation before modifying
the underlying list.