Fix string conversion of _store field
ElementTree.tostring returns a bytestring on Python 3. The default
encoding used in this case is US-ASCII. Make the default explicit
and convert the returned bytestring to a string with proper encoding.
This ensures that the _store in the final output sent to SOLR remains
the same as it is in existing versions of SIR. For the same reason,
I didn't change encoding to UTF-8.