use list instead of set
set is not json serializable by python default json module.
because we are now not using sets in the returned data dicts, the order
of items inside lists in the items can be different across test runs
and we can't use plain dict equality for asserts. instead manually
iterate the dict and use assertCountEqual for lists.