Fix macOS/spawn process failures
On macOS multiprocessing module uses `spawn` to create new processes. The arguments
passed to these processes are also pickle-d along with their dependencies. It seems
a pysolr.Solr objects are not pickl-able so instead pass the entity name to the
process and let the solr_connection be created there. Another difference between
spawn-ed and fork-ed processes is that the former do not share existing data with the
new process. As a result, the config needs to be read again by each newly spawn-ed
process.