Defer unneeded column properties
To my understanding, the `hasattr(prop, columns)` was intended to be an
indirect check for the property being a `RelationshipProperty`. Obviously,
it excluded column properties from being deffered. SQLAlchemy by default
loads all column properties eagerly.

However, there are some particularly expensive column properties, for
instance: `place_count`, `artist_count`, `label_count` on `CustomArea`.
Hence, improve the method to defer load of unneeded column properties as
well. Column properties and relationship properties are deferred using
different functions - `defer` for the former and a lazy loading strategy
for the latter.

I have chosen `raiseload` as the lazy loading strategy to ensure that
all the required data for serializing is loaded in the fewest possible
queries. This is done by listing all needed fields in the SearchField
objects. The `raiseload=True` attribute passed to `defer` has the same
rationale but for column properties. If any attributes or relations
that are not listed in the SearchField class are accessed, SQLAlchemy
will raise an error.

References:
- https://docs.sqlalchemy.org/en/20/orm/queryguide/columns.html#using-raiseload-to-prevent-deferred-column-loads
- https://docs.sqlalchemy.org/en/20/orm/queryguide/relationships.html#preventing-unwanted-lazy-loads-using-raiseload