Migrate to pyproject.toml
On the lines of other MeB python projects, migrate SIR build from setup.py to
pyproject.toml. Most of the metadata is static, and hence copied from setup.py
to pyproject.toml file and remains the same.

versioning is the major change. It is now handled automatically by `setuptools_scm`.

The other change is the addition of a `.flake8` file because `flake8` doesn't
support configuration by pyproject.toml.

References:

- https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
- https://setuptools-scm.readthedocs.io/en/latest/usage/
- https://github.com/PyCQA/flake8/issues/234/