Use readthedocs theme for SIR docs
readthedocs has started to obey the theme specified in the project's sphinx
configuration since a while. The current sphinx theme configuration of sir docs
predates RTD and was not actually not being used for as long as the docs have
been on RTD. Update the documentation to explicitly use the RTD theme for docs.